Copied the resulting HTML from the view into a component.
This should render the same structure as the original view. It still
doesn't work, because of CSRF token thing - but it's a good start. The
next steps would be to slowly transform this code to be more and more structured and not depending on in-place constants.

The HTML copied needed only small modifications, namely:

* Replaced all `class` attributes to `className` properties.
* Replaced all `for` attributes to `htmlFor` properties.

This is because both of them are reserved JavaScript words, thus can't
be used as property names.