Extracted rendering to a function. Made fix for Turbolinks.
Turbolinks needs to have their own events for handling transitions
between pages. The render part is extracted to be used in two places -
first place is a "classic" event of DOM ready and the second one is a
custom event from Turbolinks ("page:load"). This way repetition is
avoided and code works with Turbolinks.