Extracted back link and endpoint information and passed it as component's property.
This is a default way to pass constants to React components - by
including them in component's _properties_. It looks like passing an
attribute in the HTML tag.

This way you can pass whatever configuration options you want to rely on
during rendering. Note that properties can't be changed later without
re-rendering the component.