Extracted and passed CSRF token as a property.
There is a new function called `csrfToken` which is reponsible for
extracting the CSRF token from the meta tag inside your Rails layout.
Passing it as a property and replacing the corresponding input value
with it makes it work again. This is only a side-step. The final goal is
to use AJAX here, though.