Extracted the ProductForm component.
React.js is all about composing your user interface from smaller parts
called components. Components are composable just like HTML tags - but
can contain both logic and markup. This is a step to make this view more
composable and 'React'-ish. Not to mention you can re-use the product
form later since it is parametrized with method, id, class and button
label. You can use ProductForm to make editing in a seamless way.

By the way - do you see a correspondence between how view is rendered on
the Rails side (header + partial + back link) and how the structure of
your component reflects it now? It is not a coincidence!