Extracted ResourceField component.
Yet another component that can be re-used thorough Rails forms. Also
makes the purpose more clear and on the top-level there is a little more
clear what's going on.

As you may propably noticed there is a lack of validation logic and so
on. But the best part here is you can implement the validation handling
logic _once_ and all fields can benefit out of it since you've extracted
this component.