Exposed submit event to be catched by an external function.
You generally should avoid writing data processing related logic in
React components. It's way better if you provide a handler available as
property in React component so you can do anything with both event and
data which happens to be current during submitting. In this very simple
implementation you just log the received data to check if everything is
correct. `event.preventDefault()` toggles off the default behaviour of
form to be sent with a full page reload.