Created alert states. Exposed setting alert to submit handler.
So this is one of approaches how you can make the "come back" to
component after leaving it. Basically you get a function you can call
and then set state of a component.

The good part about it is that you can have a fine-grained control over
effects that such external function can have - it all depends on an
implementation of an handler you expose to an external world.

This finishes the real work here. What needs to be done is removing the
old form - you don't need it anymore.